feat(skill): redesign compress-markdown as agent-driven skill
Replace the Python scripts package with an agent-driven approach:
SKILL.md contains the compression rules the agent follows directly,
and a single validate.py provides deterministic structural verification.

Adds frontmatter, inline code, and path validation as errors. Removes
cli, detect, compress, benchmark, and package scaffolding.
