--- skills/compress-markdown/scripts/__init__.py
@@ -1,9 +1,0 @@
-"""Compress scripts.
-
-Tools to compress natural language markdown files
-into concise format to save input tokens.
-"""
-
-__all__ = ["cli", "compress", "detect", "validate"]
-
-__version__ = "1.0.0"
